JAMES-3427 Delete StoreMailboxManager and push down methods to implementations
