+++ mailbox/maildir/src/test/java/org/apache/james/mailbox/maildir/FullUserMaildirMailboxManagerTest.java
@@ -24,13 +24,12 @@ import org.apache.james.junit.TemporaryFolderExtension;
 import org.apache.james.mailbox.MailboxManagerTest;
 import org.apache.james.mailbox.SubscriptionManager;
 import org.apache.james.mailbox.events.EventBus;
-import org.apache.james.mailbox.store.StoreMailboxManager;
 import org.apache.james.mailbox.store.StoreSubscriptionManager;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-class FullUserMaildirMailboxManagerTest extends MailboxManagerTest<StoreMailboxManager> {
+class FullUserMaildirMailboxManagerTest extends MailboxManagerTest<MaildirMailboxManagerProvider.MaildirMailboxManager> {
 
     @Disabled("Maildir is using DefaultMessageId which doesn't support full feature of a messageId, which is an essential" +
         " element of the Vault")
@@ -40,10 +39,10 @@ class FullUserMaildirMailboxManagerTest extends MailboxManagerTest<StoreMailboxM
 
     @RegisterExtension
     TemporaryFolderExtension temporaryFolder = new TemporaryFolderExtension();
-    Optional<StoreMailboxManager> mailboxManager = Optional.empty();
+    Optional<MaildirMailboxManagerProvider.MaildirMailboxManager> mailboxManager = Optional.empty();
 
     @Override
-    protected StoreMailboxManager provideMailboxManager() {
+    protected MaildirMailboxManagerProvider.MaildirMailboxManager provideMailboxManager() {
         if (!mailboxManager.isPresent()) {
             mailboxManager = Optional.of(createMailboxManager());
         }
@@ -55,7 +54,7 @@ class FullUserMaildirMailboxManagerTest extends MailboxManagerTest<StoreMailboxM
         return new StoreSubscriptionManager(provideMailboxManager().getMapperFactory());
     }
 
-    private StoreMailboxManager createMailboxManager() {
+    private MaildirMailboxManagerProvider.MaildirMailboxManager createMailboxManager() {
         try {
             return MaildirMailboxManagerProvider.createMailboxManager("/%fulluser", temporaryFolder.getTemporaryFolder().getTempDir());
         } catch (Exception e) {
@@ -64,7 +63,7 @@ class FullUserMaildirMailboxManagerTest extends MailboxManagerTest<StoreMailboxM
     }
 
     @Override
-    protected EventBus retrieveEventBus(StoreMailboxManager mailboxManager) {
+    protected EventBus retrieveEventBus(MaildirMailboxManagerProvider.MaildirMailboxManager mailboxManager) {
         return mailboxManager.getEventBus();
     }
 }
