+++ server/protocols/jmap-draft/src/test/java/org/apache/james/jmap/draft/methods/GetMessagesMethodTest.java
@@ -54,6 +54,7 @@ import org.apache.james.jmap.draft.utils.JsoupHtmlTextExtractor;
 import org.apache.james.jmap.memory.projections.MemoryMessageFastViewProjection;
 import org.apache.james.mailbox.BlobManager;
 import org.apache.james.mailbox.FlagsBuilder;
+import org.apache.james.mailbox.MailboxManager;
 import org.apache.james.mailbox.MailboxSession;
 import org.apache.james.mailbox.MailboxSessionUtil;
 import org.apache.james.mailbox.MessageIdManager;
@@ -65,7 +66,6 @@ import org.apache.james.mailbox.model.ComposedMessageId;
 import org.apache.james.mailbox.model.MailboxId;
 import org.apache.james.mailbox.model.MailboxPath;
 import org.apache.james.mailbox.model.MessageId;
-import org.apache.james.mailbox.store.StoreMailboxManager;
 import org.apache.james.metrics.logger.DefaultMetricFactory;
 import org.apache.james.metrics.tests.RecordingMetricFactory;
 import org.apache.james.mime4j.message.BodyPartBuilder;
@@ -99,7 +99,7 @@ public class GetMessagesMethodTest {
     private org.apache.james.mime4j.dom.Message messageContent1;
     private org.apache.james.mime4j.dom.Message messageContent2;
     private org.apache.james.mime4j.dom.Message messageContent3;
-    private StoreMailboxManager mailboxManager;
+    private MailboxManager mailboxManager;
     private GetMessagesMethod testee;
     private MailboxSession session;
     private MailboxPath inboxPath;
